Update logic for market event parameter and rename gaUserId to as24VisitorId
